refactor(jsx-directive): expose plugin to support browser environments
