+++ .changeset/clean-rabbits-tickle.md
@@ -1,0 +1,6 @@
+---
+"@vue-macros/jsx-directive": patch
+---
+
+expose plugin to support browser environments
+
